Separate out base classes and get srw tests working locally
